+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/record/HistoryDataProviderTest.kt
@@ -31,7 +31,6 @@ import com.mapbox.test.dsl.TestCase
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.spyk
-import org.junit.Ignore
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach